refactor: moved enums from types to constants folder
